Moving plugin monitor into the plugin module
